+++ src/main/java/io/contek/invoker/deribit/api/common/_LightOrder.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.deribit.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _LightOrder {
@@ -9,25 +8,25 @@ public class _LightOrder {
   public String order_state;
 //  public Double max_show;
 //  public Boolean api;
-  public BigDecimal amount;
+  public Double amount;
 //  public Boolean web;
   public String instrument_name;
 //  public String advanced;
 //  public Boolean triggered;
 //  public Boolean block_trade;
 //  public String original_order_type;
-  public BigDecimal price;
+  public Double price;
   public String time_in_force;
 //  public Boolean auto_replaced;
 //  public String stop_order_id;
   public long last_update_timestamp;
   public Boolean post_only;
 //  public Boolean replaced;
-  public BigDecimal filled_amount;
+  public Double filled_amount;
 //  public Double average_price;
   public String order_id;
   public Boolean reduce_only;
-//  public BigDecimal commission;
+//  public Double commission;
 //  public String app_name;
 //  public Double stop_price;
   public String label;
