+++ src/main/java/io/contek/invoker/deribit/api/common/_LightTicker.java
@@ -1,22 +1,21 @@
 package io.contek.invoker.deribit.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _LightTicker {
 
-  public BigDecimal best_ask_amount;
-  public BigDecimal best_ask_price;
-  public BigDecimal best_bid_amount;
-  public BigDecimal best_bid_price;
-  public BigDecimal current_funding;
-  public BigDecimal funding_8h;
+  public Double best_ask_amount;
+  public Double best_ask_price;
+  public Double best_bid_amount;
+  public Double best_bid_price;
+  public Double current_funding;
+  public Double funding_8h;
 //  public _Greek greeks;
 //  public Double index_price;
   public String instrument_name;
 //  public Double interest_rate;
-  public BigDecimal last_price;
+  public Double last_price;
 //  public Double mark_price;
 //  public Double max_price;
 //  public Double min_price;
