+++ src/main/java/io/contek/invoker/deribit/api/common/_OrderBookLevel.java
@@ -1,9 +1,8 @@
 package io.contek.invoker.deribit.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 
 @NotThreadSafe
-public class _OrderBookLevel extends ArrayList<BigDecimal> {
+public class _OrderBookLevel extends ArrayList<Double> {
 }
