+++ src/main/java/io/contek/invoker/deribit/api/common/_Stats.java
@@ -1,11 +1,9 @@
 package io.contek.invoker.deribit.api.common;
 
-import java.math.BigDecimal;
-
 public class _Stats {
   public Double high;
   public Double low;
   public Double price_change;
-  public BigDecimal volume;
-  public BigDecimal volume_usd;
+  public Double volume;
+  public Double volume_usd;
 }
