+++ src/main/java/io/contek/invoker/deribit/api/websocket/market/TradesChannel.java
@@ -7,7 +7,6 @@ import io.contek.invoker.deribit.api.websocket.common.WebSocketSingleChannelMess
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 import static io.contek.invoker.deribit.api.websocket.common.constants.WebSocketChannelKeys._trades;
@@ -44,7 +43,7 @@ public final class TradesChannel
   @NotThreadSafe
   public static final class Data {
 
-    public BigDecimal amount;
+    public Double amount;
     public String block_trade_id;
     public String direction;
     public double index_price;
@@ -52,7 +51,7 @@ public final class TradesChannel
     public double iv;
     public String liquidation;
     public double mark_price;
-    public BigDecimal price;
+    public Double price;
     public int tick_direction;
     public long timestamp;
     public String trade_id;
